jd_fruit.js更新支持京东双账号,Node.js用户支持N个京东账号 功能更新:为好友浇水 action 添加 jd_fruit.yml
